Create Polyjuice Potion.py
Hermione wants to make a polyjuice potion again to find out if Draco is onto something. Now that she became an auror, she is kind of a pro in making one, and she has her own polyjuice essence. She only needs to mix it with castor oil, to make a perfect polyjuice. But there is a problem; the oil and essence are stored in bottles in her basement and whenever she opens a bottle she has to use it all. Therefore, during making the potion, she can add one bottle in each step.

Given the percentage of the castor oil in the whole potion, find out the minimum number of bottles (in total) she used.

Input Format:
First line contains n as the number of potions formed by different percentages of ingredients.

The following n lines contain 1 integer each, ai, the percentage of castor oil in the polyjuice.

Constraints:
1≤n≤100
1≤ai≤100
Output Format:
For each line: minimum total number of bottles
Sample Input:
Copy
3
5
20
100
Sample Output:
Copy
20
5
1
Explanation:
Copy
For the first potion:

Percentage of the castor oil in the first potion should be 5, and the rest of the polyjuice should be essence, so the percentage ratio of oil/essence = 5/95 can be simplified into 1/19 and in total there would be 20 bottles used.